pyproject.toml
[project]
name = "test-strip-analyzer"
version = "0.1.0"
description = "Analyze 2-in-1 test strip images to determine positive/negative results"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core image processing
    "numpy>=1.24",
    "opencv-python>=4.8",
    "pillow>=10.0",
    "scikit-image>=0.21",
    # ML
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    # Data handling
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Serving
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
    # Utilities
    "joblib>=1.3",
    "pyyaml>=6.0",
    "matplotlib>=3.10.8",
]

[project.optional-dependencies]
cnn = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
]
bgremoval = [
    "transparent-background>=1.0",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "ruff>=0.1",
    "ipykernel>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["training", "inference", "serving"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
